Delete the core and restore functionality
This version also works. Duplicate functions in the core have been
deleted and now all functions are split into their respective files.
Note that in order to correct this code to run properly, I had to remove
several circular deps and add a new file, as well as make many smaller
changes. Do not go back to the previous commit for any reason